Only use PriorityAsStringOrder for display
Any value of PriorityAsStringOrder which was different from the computed
set (by re-ordering or substraction of levels) would also cause the
levels for display to be incorrect.  Fix the algorithm to be independent
of the PriorityAsStringOrder setting, and update the documentation to
make it clearer that most use cases do not require setting it.